Fix millis() bug in LightService, fix erroneous use of unit8_t, remove a 0 from FADE_SPEED definition, change bri_multiplier from uint8_t to float
LightService was previously not rollover safe in its use of millis() - this has been resolved.  I also replaced the int argument of rise_set_to_time with uint8_t when in fact it has to be an int, as rise_set could be as high as 1439.

I had also used a uint8_t for the bri_multiplayer member of the LightController class, when in fact this should be a float, or the light will never turn on.